backend/requirements.txt
asgiref==3.9.1
attrs==25.3.0
django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.4.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2024.10.1
python-dotenv==1.0.1
PyJWT==2.10.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.21.0
sqlparse==0.5.3
uritemplate==4.2.0
psycopg2-binary==2.9.10
